cloudrun/requirements.txt
flask
scikit-learn
pandas
numpy
google-cloud-storage
joblib
matplotlib
